requirements/testing.txt
#
# This file is autogenerated by pip-compile with python 3.6
# To update, run:
#
#    pip-compile requirements/testing.in
#
attrs==21.4.0
    # via pytest
backports.zoneinfo==0.2.1
    # via
    #   -r requirements/production.txt
    #   pytz-deprecation-shim
    #   tzlocal
beautifulsoup4==4.11.1
    # via
    #   -r requirements/production.txt
    #   o365
certifi==2022.6.15
    # via
    #   -r requirements/production.txt
    #   requests
charset-normalizer==2.0.12
    # via
    #   -r requirements/production.txt
    #   requests
idna==3.3
    # via
    #   -r requirements/production.txt
    #   requests
importlib-metadata==4.8.3
    # via
    #   pluggy
    #   pytest
importlib-resources==5.4.0
    # via
    #   -r requirements/production.txt
    #   backports.zoneinfo
iniconfig==1.1.1
    # via pytest
o365==2.0.19
    # via -r requirements/production.txt
oauthlib==3.2.0
    # via
    #   -r requirements/production.txt
    #   requests-oauthlib
packaging==21.3
    # via pytest
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest
pyparsing==3.0.9
    # via packaging
pytest==7.0.1
    # via -r requirements/testing.in
python-dateutil==2.8.2
    # via
    #   -r requirements/production.txt
    #   o365
pytz==2022.1
    # via
    #   -r requirements/production.txt
    #   o365
pytz-deprecation-shim==0.1.0.post0
    # via
    #   -r requirements/production.txt
    #   tzlocal
pyyaml==6.0
    # via -r requirements/production.txt
requests==2.27.1
    # via
    #   -r requirements/production.txt
    #   o365
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via
    #   -r requirements/production.txt
    #   o365
six==1.16.0
    # via
    #   -r requirements/production.txt
    #   python-dateutil
soupsieve==2.3.2.post1
    # via
    #   -r requirements/production.txt
    #   beautifulsoup4
stringcase==1.2.0
    # via
    #   -r requirements/production.txt
    #   o365
tomli==1.2.3
    # via pytest
typing-extensions==4.1.1
    # via importlib-metadata
tzdata==2022.1
    # via
    #   -r requirements/production.txt
    #   pytz-deprecation-shim
tzlocal==4.2
    # via
    #   -r requirements/production.txt
    #   o365
urllib3==1.26.10
    # via
    #   -r requirements/production.txt
    #   requests
zipp==3.6.0
    # via
    #   -r requirements/production.txt
    #   importlib-metadata
    #   importlib-resources
